Fix subclassing mechanism; switch from `es6-shim` to `core-js`.
Fix our subclass code to follow the latest ES6 specification, which
requires the use of the `new` operator when creating subclass instances.

Switch from `es6-shim` to `core-js` for a default `Promise` implementation,
which allows us to apply a narrow shim for `Promise` instead of requiring
all of ES6 to be patched.  `core-js` also allows us to completely
avoid modifying globals.