Fixes to improve interoperability with native promises.
Native promise implementations (ie, in iojs) don't always recognize
the "arguments" object as an iterable, even though the ES6 spec says
it is one.  Don't be a pedant, just Array#slice it so that we work
nicely together.

Similarly, native promise implementations use a hidden
[[PromiseConstructor]] field which there is no way to access
at the moment (since the engines don't implement the rest of the
ES6 subclassing semantics).  Work around this by patching the
Promise#resolve method.